Got 1 tile to show at least...

+++ GameOfSomeSort/src/Item.java
@@ -2,10 +2,10 @@ public class Item {
 
     final int DEFAULT_VALUE = 100;
     final int DEFAULT_SPACE = 1;
-    final int DEFAULT_RARITY = 1;
+    final double DEFAULT_RARITY = 1.0;
 
     private int value = this.DEFAULT_VALUE;
     private int space = this.DEFAULT_SPACE;
-    private int rarity = this.DEFAULT_RARITY;
+    private double rarity = this.DEFAULT_RARITY;
 
 }
